# internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Chain struct {
	ID             pgtype.UUID
	Name           string
	ChainID        int32
	RpcUrl         string
	NativeCurrency string
	ExplorerUrl    pgtype.Text
	Status         string
	CreatedAt      pgtype.Timestamp
	UpdatedAt      pgtype.Timestamp
}

type Token struct {
	ID              pgtype.UUID
	ChainID         pgtype.UUID
	ContractAddress string
	Name            string
	Symbol          string
	Decimals        int32
	LogoUrl         pgtype.Text
	Type            string
	Status          string
	CreatedAt       pgtype.Timestamp
	UpdatedAt       pgtype.Timestamp
}

type Transaction struct {
	ID          pgtype.UUID
	ChainID     int32
	FromAddress string
	ToAddress   string
	TxHash      string
	CreatedAt   pgtype.Timestamp
	UpdatedAt   pgtype.Timestamp
}

type User struct {
	ID           pgtype.UUID
	Email        string
	PasswordHash string
	Status       string
	CreatedAt    pgtype.Timestamp
	UpdatedAt    pgtype.Timestamp
}

type Wallet struct {
	ID                  pgtype.UUID
	UserID              pgtype.UUID
	Address             string
	EncryptedPrivateKey []byte
	Name                pgtype.Text
	Status              string
	CreatedAt           pgtype.Timestamp
	UpdatedAt           pgtype.Timestamp
}